% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BP_narrow}
\alias{BP_narrow}
\alias{BP_wide}
\title{Blood pressure readings
Wide and narrow representations of blood pressure data.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
BP_narrow

BP_wide
}
\description{
Blood pressure readings
Wide and narrow representations of blood pressure data.
}
\keyword{datasets}
