% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CholeraDeaths}
\alias{CholeraDeaths}
\alias{CholeraPumps}
\title{Spatial data on Deaths and Pumps from 1854 London cholera outbreak}
\format{An object of class \code{\link[sp]{SpatialPointsDataFrame}} 
with 250 rows and 2 columns.}
\source{
\url{http://blog.rtwilson.com/john-snows-cholera-data-in-more-formats/}
}
\usage{
CholeraDeaths

CholeraPumps
}
\description{
Spatial data on Deaths and Pumps from 1854 London cholera outbreak
}
\keyword{datasets}

