% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountryData}
\alias{CountryData}
\alias{CountryData2}
\title{Many variables on countries from the CIA factbook, 2014.}
\format{A data frame on the following 76 variables for each of the Countries
in the World. (256 countries are given.)  \describe{
\item{country}{Name of the country.} \item{area}{area (sq
km), 2147} \item{pop}{number of people, 2119}
\item{growth}{growth rate per year (\%),2002}
\item{birth}{birth rate (#/1000), 2054} \item{death}{death
rate (#/1000), 2066} \item{netMigr}{net migration per year (#/1000).
Difference between incoming and outgoing. 2112}
\item{maternal}{maternal deaths per 100000 live births. 2223}
\item{infant}{infant deaths per 1000 live births. 2091}
\item{life}{life expectancy (years), 2102}
\item{fert}{children born/woman (#/person), 2127}
\item{health}{health spending (\% of GDP), 2225}
\item{HIVrate}{people with HIV (\% of population), 2155}
\item{HIVpeople}{people with HIV (#), 2156}
\item{HIVdeath}{deaths from HIV (#/year), 2157}
\item{obesity}{fraction of population considered obese (\%), 2228}
\item{underweight}{fraction of population considered underweight
(\%), 2224} \item{educ}{education spending (\% of GDP), 2206}
\item{unemploymentYouth}{youth unemployment (\%), 2229}
\item{GDP}{Gross Domestic Product ($), 2001}
\item{GDPgrowth}{growth in GDP (\%/year), 2003}
\item{GDPcapita}{GDP per capita ($/person), 2004}
\item{saving}{yearly money savings (\% of GDP), 2260}
\item{indProd}{Industrial production growth rate (\%), 2089}
\item{labor}{Labor force (people), 2095}
\item{unemployment}{Unemployment rate (\%), 2129}
\item{family}{Distribution of family income - Gini index (Gini
index), 2172} \item{tax}{Taxes and other revenues (\% of GDP), 2221}
\item{budget}{Budget surplus (+) or deficit (-) (\% of GDP), 2222}
\item{debt}{Public debt (\% of GDP), 2186}
\item{inflation}{Inflation rate (consumer prices) (\%), 2092}
\item{discount}{Central bank discount rate (\%), 2207}
\item{lending}{Commercial bank prime lending rate (\%), 2208}
\item{narrow}{Stock of narrow money ($), 2214}
\item{broad}{Stock of broad money ($), 2215}
\item{credit}{Stock of domestic credit ($), 2211}
\item{shares}{Market value of publicly traded shares ($), 2200}
\item{balance}{Current account balance ($), 2187}
\item{exports}{Exports ($), 2078} \item{imports}{Imports
($), 2087} \item{gold}{Reserves of foreign exchange and gold ($),
2188} \item{externalDebt}{Debt - external ($), 2079}
\item{homeStock}{Stock of direct foreign investment - at home ($),
2198} \item{abroadStock}{Stock of direct foreign investment - abroad
($), 2199} \item{elecProd}{Electricity - production (kWh), 2232}
\item{elecCons}{Electricity - consumption (kWh), 2233}
\item{elecExp}{Electricity - exports (kWh), 2234}
\item{elecImp}{Electricity - imports (kWh), 2235}
\item{elecCap}{Electricity - installed generating capacity (kW),
2236} \item{elecFossil}{Electricity - from fossil fuels (\% of total
installed capacity), 2237} \item{elecNuc}{Electricity - from nuclear
fuels (\% of total installed capacity), 2239}
\item{elecHydro}{Electricity - from hydroelectric plants (\% of
total installed capacity), 2238} \item{elecRenew}{Electricity - from
other renewable sources (\% of total installed capacity), 2240}
\item{oilProd}{Crude oil - production (bbl/day), 2241}
\item{oilExp}{Crude oil - exports (bbl/day), 2242}
\item{oilImp}{Crude oil - imports (bbl/day), 2243}
\item{oilRes}{Crude oil - proved reserves (bbl), 2244}
\item{petroProd}{Refined petroleum products - production (bbl/day),
2245} \item{petroCons}{Refined petroleum products - consumption
(bbl/day), 2246} \item{petroExp}{Refined petroleum products -
exports (bbl/day), 2247} \item{petroImp}{Refined petroleum products
- imports (bbl/day), 2248} \item{gasProd}{Natural gas - production
(cu M), 2249} \item{gasCons}{Natural gas - consumption (cu M), 2250}
\item{gasExp}{Natural gas - exports (cu M), 2251}
\item{gasImp}{Natural gas - imports (cu M), 2252}
\item{gasRes}{Natural gas - proved reserves (cu M), 2253}
\item{mainlines}{Telephones - main lines in use (mainlines in use),
2150} \item{cell}{Telephones - mobile cellular (cellphones), 2151}
\item{netHosts}{Internet hosts (#), 2184}
\item{netUsers}{Internet users (#), 2153}
\item{airports}{Airports (#), 2053} \item{railways}{Railways
(km), 2121} \item{roadways}{Roadways (km), 2085}
\item{waterways}{Waterways (km), 2093}
\item{marine}{Merchant marine (# of ships), 2108}
\item{military}{Military expenditures (\% of GDP), 2034} }}
\source{
From the CIA World Factbook,
\url{https://www.cia.gov/library/publications/the-world-factbook/} For
package maintainers: The lines to construct the file are in
\code{inst/datafixes.R}.
}
\usage{
CountryData

CountryData2
}
\description{
The CIA Factbook has geographic, demographic, and economic data on a
country-by-country basis.  In the description of the variables, the 4-digit
number indicates the code used to specify that variable on the data and
documentation web site.  For instance,
\url{https://www.cia.gov/library/publications/the-world-factbook/fields/2153.html}
contains documentation for variable code 2153, network users.

A modified version of CountryData
}
\details{
These data were collected using the \code{CIAdata()} function.

THE DOCUMENTATION NEEDS COMPLETING.  See the output of \code{CIAdata() and
Google "CIA Factbook NNNN"}, where NNNN is the 4-digit code used for
indexing the data at the CIA.
}
\examples{

data(CountryData)
## maybe str(CountryData) ; plot(CountryData) ...

}
\keyword{datasets}

