% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountryData2}
\alias{CountryData2}
\title{A modified version of CountryData}
\format{\preformatted{'data.frame':	111 obs. of  6 variables:
 $ country  : chr  "Albania" "Algeria" "Angola" "Argentina" ...
 $ gdp      : num  9383 7335 6905 17920 44354 ...
 $ educ     : num  3.3 4.3 3.5 6.3 5.6 5.9 2.4 2.6 2.2 5.6 ...
 $ roadways : num  0.63 0.05 0.04 0.08 0.11 1.48 0.61 5.42 0.15 3.72 ...
 $ net_users: Factor w/ 5 levels ">0\%",">5\%",">15\%",..: 4 2 1 3 5 5 3 3 1 5 ...
  ..- attr(*, "names")= chr  "4" "2" "1" "3" ...
 $ oilProd2 : Factor w/ 6 levels "<1k","<10k","<100k",..: 3 5 5 4 4 3 4 3 2 2 ...
 - attr(*, "na.action")=Class 'omit'  Named int [1:145] 1 2 5 6 8 9 10 12 13 14 ...
  .. ..- attr(*, "names")= chr [1:145] "1" "2" "5" "6" ...
}}
\usage{
CountryData2
}
\description{
A modified version of CountryData
}
\keyword{datasets}

