% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cuisines}
\alias{Cuisines}
\title{Cuisines}
\format{{
A data frame with 531420 observations on the following 15 variables.
\describe{
 \item{\code{cuisine_code}}{a character vector}
 \item{\code{cuisine_description}}{a character vector}
}
}}
\usage{
Cuisines
}
\description{
Cuisines
}
\keyword{datasets}

