% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Elections}
\alias{Elections}
\title{Election Statistics}
\format{
An object of class \link[tibble:tbl_df-class]{tibble::tbl_df} with 117 rows and 13 columns.
\describe{
\item{Ward}{Name of the country}
\item{Precinct}{number of people, 2119}
\item{Registered.Voters.at.7am}{area (sq km), 2147}
\item{Voters.Registering.at.Polls}{Crude oil - production (bbl/day), 2241}
\item{gdp}{Gross Domestic Product per capita ($/person), 2001}
\item{educ}{education spending (\% of GDP), 2206}
\item{roadways}{Roadways per unit area (km/sq km), 2085}
\item{net_users}{Fraction of Internet users (\% of population), 2153}
}
}
\usage{
Elections
}
\description{
Election Statistics
}
\keyword{datasets}
