% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electionStatistics}
\alias{electionStatistics}
\title{election statistics}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 117 rows and 13 columns.}
\usage{
electionStatistics
}
\description{
election statistics
}
\keyword{datasets}

