% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Emails_test}
\alias{Emails_test}
\title{Email Test}
\format{
A data frame with 1,382 rows and 3 variables:
\describe{
  \item{subjectline}{String}
  \item{type}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
Emails_test
}
\description{
This dataset is a random sample of 20\% of the emails data.
}
\keyword{datasets}
