% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Emails_train}
\alias{Emails_train}
\alias{Emails_test}
\title{Email Train}
\format{
A data frame with 5,526 rows and 3 variables:
\describe{
  \item{ids}{an integer vector}
  \item{subjectline}{a character vector}
  \item{type}{a character vector}
}

A data frame with 1,382 rows and 3 variables:
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
Emails_train

Emails_test
}
\description{
The training dataset includes a set of email subject lines used for classification
of whether the message is spam (unsolicited commercial content) or not.
Many subject lines include subject matter inappropriate for classroom use. 
Given the volume of headlines containing such language 
(especially for \code{spam == TRUE}), user discretion is advised.
This dataset is a random sample of 80\% of the emails data.

The testing dataset is a random sample of 20\% of the emails data.
}
\examples{
nrow(Emails_train)
nrow(Emails_test)
}
\keyword{datasets}
