% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCI60.R
\name{etl_NCI60}
\alias{etl_NCI60}
\title{Load the NCI60 data from GitHub}
\usage{
etl_NCI60()
}
\value{
A \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Load the NCI60 data from GitHub
}
\examples{

\donttest{
# The file is 5.0 MB
NCI60 <- etl_NCI60()
}
}
