% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Headlines_test}
\alias{Headlines_test}
\title{Headlines Test}
\format{
A data frame with 4589 rows and 3 variables:
\describe{
  \item{title}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
Headlines_test
}
\description{
This dataset is a random sample of the remaining 20\% of the observations
not found in the training set.
}
\keyword{datasets}
