% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Macbeth_raw}
\alias{Macbeth_raw}
\title{Text of Macbeth}
\format{
A character vector of length 1
}
\source{
Project Gutenberg, \url{https://www.gutenberg.org/ebooks/1129/}
}
\usage{
Macbeth_raw
}
\description{
The entire text of Macbeth, stored in a character vector of length 1.
}
\keyword{datasets}
