% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{mdsr_table}
\alias{mdsr_table}
\alias{mdsr_sql_explain_table}
\alias{mdsr_sql_keys_table}
\title{Custom table output}
\usage{
mdsr_table(x, ...)

mdsr_sql_explain_table(x, ...)

mdsr_sql_keys_table(x, ...)
}
\arguments{
\item{x}{A data.frame}

\item{...}{arguments passed to \code{\link[kableExtra:kbl]{kableExtra::kbl()}}}
}
\description{
Custom table output
}
\examples{
mdsr_table(faithful)
}
