% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ordway_birds}
\alias{ordway_birds}
\title{Birds captured and released at Ordway, complete and uncleaned}
\format{
A data frame with 15,829 observations on the bird's species, size,
date found, and band number.
\describe{
\item{bogus}{a character vector}
\item{Timestamp}{Timestamp indicates when the data were entered into an electronic record,
not anything about the bird being described}
\item{Year}{a character vector}
\item{Day}{a character vector}
\item{Month}{a character vector}
\item{CaptureTime}{a character vector}
\item{SpeciesName}{a character vector}
\item{Sex}{a character vector}
\item{Age}{a character vector}
\item{BandNumber}{a character vector}
\item{TrapID}{a character vector}
\item{Weather}{a character vector}
\item{BandingReport}{a character vector}
\item{RecaptureYN}{a character vector}
\item{RecaptureMonth}{a character vector}
\item{RecaptureDay}{a character vector}
\item{Condition}{a character vector}
\item{Release}{a character vector}
\item{Comments}{a character vector}
\item{DataEntryPerson}{a character vector}
\item{Weight}{a character vector}
\item{WingChord}{a character vector}
\item{Temperature}{a character vector}
\item{RecaptureOriginal}{a character vector}
\item{RecapturePrevious}{a character vector}
\item{TailLength}{a character vector}
}

Timestamp indicates when the data were entered into an electronic record,
not anything about the bird being described.
}
\source{
Jerald Dosch, Dept. of Biology, Macalester College: the manager of
the Study Area.
}
\usage{
ordway_birds
}
\description{
The historical record of birds captured and released at the Katharine Ordway
Natural History Study Area, a 278-acre preserve in Inver Grove Heights,
Minnesota, owned and managed by Macalester College.
}
\details{
There are many extraneous levels of variables such as species.  Part of the
purpose of this data set is to teach about data cleaning.
}
\examples{

ordway_birds

}
\references{
\url{https://www.macalester.edu/ordway/}
}
\keyword{datasets}
