% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrdwayBirds}
\alias{OrdwayBirds}
\title{Birds captured and released at Ordway, complete and uncleaned}
\format{A data frame with 15,829 observations on the bird's species, size,
date found, and band number.

Timestamp indicates when the data were entered into an electronic record,
not anything about the bird being described.}
\source{
Jerald Dosch, Dept. of Biology, Macalester College: the manager of
the Study Area.
}
\usage{
OrdwayBirds
}
\description{
The historical record of birds captured and released at the Katharine Ordway
Natural History Study Area, a 278-acre preserve in Inver Grove Heights,
Minnesota, owned and managed by Macalester College.
}
\details{
There are many extraneous levels of variables such as species.  Part of the
purpose of this data set is to teach about data cleaning.

A translation table for cleaning the species names is in
\code{\link[DataComputing]{OrdwaySpecies}}.
}
\examples{

data(OrdwayBirds)

}
\keyword{datasets}

