% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saratoga_houses}
\alias{saratoga_houses}
\alias{saratoga_codes}
\title{Saratoga Houses}
\format{
A tibble with 1728 rows and 16 variables:
\describe{
\item{price}{},
\item{lot_size}{},
\item{waterfront}{},
\item{age}{},
\item{land_value}{},
\item{construction}{},
\item{air_cond}{},
\item{fuel}{},
\item{heat}{},
\item{sewer}{},
\item{living_area}{},
\item{pct_college}{},
\item{bedrooms}{},
\item{fireplaces}{},
\item{bathrooms}{},
\item{rooms}{}
}
@examples
saratoga_houses

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 3 columns.
}
\usage{
saratoga_houses

saratoga_codes
}
\description{
Saratoga Houses
}
\keyword{datasets}
