% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{skim}
\alias{skim}
\title{Custom skimmer}
\usage{
skim(data, ...)
}
\arguments{
\item{data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\description{
Custom skimmer
}
\examples{
skim(faithful)
}
