% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ViolationCodes}
\alias{ViolationCodes}
\title{NYC Restaurant Health Violation Codes}
\format{A data frame with 174 observations on the following 3 variables.
\describe{
   \item{\code{violation_code}}{a factor with many levels}
   \item{\code{critical_flag}}{is violation critical: a factor with levels \code{N} \code{Y}}
   \item{\code{violation_description}}{violation description}
 }}
\usage{
ViolationCodes
}
\description{
NYC Restaurant Health Violation Codes
}
\keyword{datasets}

