% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_cities}
\alias{world_cities}
\title{Cities and their populations}
\format{
A data frame with 4,428 observations on the following 10 variables.
\describe{ 
\item{geoname_id}{integer id of record in geonames database}
\item{name}{name of geographical point in plain ascii characters} 
\item{latitude}{latitude in decimal degrees (wgs84)} 
\item{longitude}{longitude in decimal degrees (wgs84)}
\item{country}{ISO-3166 2-letter country code}
\item{country_region}{fipscode}
\item{population}{Population} 
\item{timezone}{the iana timezone id} 
\item{modification_date}{date of last modification} }
}
\source{
GeoNames: \url{http://download.geonames.org/export/dump/}
}
\usage{
world_cities
}
\description{
A list of cities
}
\examples{

world_cities

}
\keyword{datasets}
