% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldCities}
\alias{WorldCities}
\title{Cities and their populations}
\format{A data frame with 23018 observations on the following 10 variables.
\describe{ \item{list("code")}{The ISO (?) city code}
\item{list("name")}{Name of the city} \item{list("latitude")}{location in
degrees} \item{list("longitude")}{location in degrees}
\item{list("country")}{Two letter country code}
\item{list("countryRegion")}{A numerical region}
\item{list("population")}{Population} \item{list("regionCode")}{ISO (?)
Code} \item{list("region")}{Name of the region} \item{list("date")}{Date
estimate made} }}
\source{
In Draft: Somewhere on the Internet.  We need a proper source.
}
\usage{
WorldCities
}
\description{
A list of cities
}
\examples{

data(WorldCities)
## maybe str(WorldCities) ; plot(WorldCities) ...

}
\references{
In Draft: We need a proper reference
}
\keyword{datasets}

