% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cherry}
\alias{Cherry}
\title{Cherry Blossom runs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41248 rows and 8 columns.}
\usage{
Cherry
}
\description{
Cherry Blossom runs
}
\keyword{datasets}

