% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CholeraDeaths}
\alias{CholeraDeaths}
\alias{CholeraPumps}
\title{Deaths and Pumps from 1854 London cholera outbreak}
\format{
An object of class \code{\link[sf]{sf}} 
whose data attribute has 250 rows and 2 columns.

An object of class \code{\link[sf]{sf}}.
}
\source{
\url{https://blog.rtwilson.com/john-snows-cholera-data-in-more-formats/}
}
\usage{
CholeraDeaths

CholeraPumps
}
\description{
Deaths and Pumps from 1854 London cholera outbreak
}
\details{
Both spatial objects are projected in EPSG:27700, aka the British
National Grid.
}
\examples{
if (require(sf)) {
  plot(st_geometry(CholeraDeaths))
}
}
\keyword{datasets}
