% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountryData2}
\alias{CountryData2}
\title{A modified version of CountryData}
\format{An object of class \code{data.frame} with 111 rows and 6 columns.}
\usage{
CountryData2
}
\description{
A modified version of CountryData
}
\keyword{datasets}

