% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Elections}
\alias{Elections}
\title{Election Statistics from the 2013 Minneapolis Mayoral Election}
\format{
An object of class \link[tibble:tbl_df-class]{tibble::tbl_df} with 117 rows and 13 columns.
\describe{
\item{Ward}{Number of the ward}
\item{Precinct}{Number of the precinct}
\item{Registered Voters at 7am}{Number of registered votes as of 7 am}
\item{Voters Registering at Polls}{Number of voters registering at the polls}
\item{Voters Registering by Absentee}{Number of voters registering by absentee}
\item{Total Registrations}{Total number of registered voters}
\item{Voters at Polls}{Number of voters at the polls}
\item{Absentee Voters}{Number of absentee voters}
\item{Total Ballots Cast}{Number of total ballots cast}
\item{Total Turnout}{Total number of voters turning out}
\item{Percentage Absentee}{Percentage of absentee voters}
\item{\% Registered to Total (Election Day)}{Percentage of voters relative to total number of people}
\item{Spoiled Ballots}{Number of spolied ballots}
}
}
\source{
\url{https://vote.minneapolismn.gov/results-data/election-results/2013/mayor/}
}
\usage{
Elections
}
\description{
Election Statistics from the 2013 Minneapolis Mayoral Election
}
\keyword{datasets}
