% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electionStatistics}
\alias{electionStatistics}
\title{election statistics}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	117 obs. of  13 variables:
 $ Ward                                 : int  1 1 1 1 1 1 1 1 1 2 ...
 $ Precinct                             : Factor w/ 16 levels "1","10","1C",..: 1 4 6 8 10 12 14 15 16 1 ...
 $ Registered.Voters.at.7am             : int  1878 2769 2337 2139 1875 2258 1847 1332 2401 2820 ...
 $ Voters.Registering.at.Polls          : int  25 43 40 24 31 69 47 43 42 62 ...
 $ Voters.Registering.by.Absentee       : int  3 1 0 5 0 0 0 0 0 1 ...
 $ Total.Registrations                  : int  28 44 40 29 31 69 47 43 42 63 ...
 $ Voters.at.Polls                      : int  492 836 905 768 683 739 291 415 596 1011 ...
 $ Absentee.Voters                      : int  27 56 19 26 31 20 8 5 25 39 ...
 $ Total.Ballots.Cast                   : int  519 892 924 794 714 759 299 420 621 1050 ...
 $ Total.Turnout                        : num  0.272 0.317 0.389 0.366 0.375 ...
 $ Percentage.Absentee                  : num  0.052 0.0628 0.0206 0.0327 0.0434 0.0264 0.0268 0.0119 0.0403 0.0371 ...
 $ X..Registered.to.Total..Election.Day.: num  0.0508 0.0514 0.0442 0.0313 0.0454 ...
 $ Spoiled.Ballots                      : int  14 22 34 19 14 32 17 22 15 42 ...
}}
\usage{
electionStatistics
}
\description{
election statistics
}
\keyword{datasets}

