% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Emails_train}
\alias{Emails_train}
\title{Email Train}
\format{
A data frame with 5,526 rows and 3 variables:
\describe{
  \item{subjectline}{String}
  \item{type}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{http://www.rdatasciencecases.org/Spam}
}
\usage{
Emails_train
}
\description{
This dataset includes a set of email subject lines used for classification
of whether the message is spam (unsolicited commercial content) or not.
Many subject lines include  subject matter innapropriate for classroom use. 
Given the volume of headlines containing such language 
(especially for spam == TRUE), user discretion is advised.
This dataset is a random sample of 80\% of the emails data.
}
\keyword{datasets}
