% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Headlines_train}
\alias{Headlines_train}
\title{Headlines Train}
\format{
A data frame with 18360 rows and 3 variables:
\describe{
  \item{title}{String}
  \item{clickbait}{Boolean}
  \item{ids}{Integer}
}
}
\source{
\url{https://github.com/bhargaviparanjape/clickbait}
}
\usage{
Headlines_train
}
\description{
This data comes from Chakraborty et. al., which combines headlines from
a variety of news and clickbait sources. Some headlines contain
subject matter innapropriate for classroom use. Given the volume of headlines
containing such language (especially for clickbait == TRUE), this filtering
might not catch all problematic headlines. User discretion is advised.
This dataset is a random sample of approximately 80\% of the observations
from the original dataset.
}
\keyword{datasets}
