% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{macros}
\alias{macros}
\alias{func}
\alias{sql_func}
\alias{sql_word}
\alias{argument}
\alias{variable}
\alias{pkg}
\alias{mdsr_data}
\alias{vocab}
\alias{index_entry}
\title{Replacements for LaTeX macros}
\usage{
func(x)

sql_func(x)

sql_word(x)

argument(x)

variable(x)

pkg(x)

mdsr_data(x)

vocab(x)

index_entry(index_label = "subject", x)
}
\arguments{
\item{x}{text to wrap in macro}

\item{index_label}{the name of the index}
}
\description{
Replacements for LaTeX macros
}
