% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MedicareCharges}
\alias{MedicareCharges}
\title{Charges to and Payments from Medicare}
\format{A data frame with 163065 observations on the following 5 variables.
\describe{ \item{drg}{Code for the Direct Recovery Group: a
character string that looks like a number.} \item{idProvider}{a
numeric vector: it should be a factor.} \item{totalDischarges}{a
numeric vector} \item{aveCharges}{a numeric vector}
\item{avePayments}{a numeric vector} }}
\source{
Data from the Centers for Medicare and Medicaid Services.
See \url{http://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/Inpatient.html}.
}
\usage{
MedicareCharges
}
\description{
These data for 2011, released in May 2013, describe how much hospitals
charged Medicare for various inpatient procedures, how many were performed,
and how much Medicare actually paid.
}
\details{
These data are part of a set with \code{DirectRecoveryGroups}, which gives a
description of the medical procedure associated with each DRG, and
\code{MedicareProviders}, which translates \code{idProvider} into a name,
address, state, Zip, etc..
}
\examples{

data(MedicareCharges)

}
\keyword{datasets}

