% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OrdwayBirds}
\alias{OrdwayBirds}
\title{Birds captured and released at Ordway, complete and uncleaned}
\format{A data frame with 15,829 observations on the bird's species, size,
date found, and band number.
\describe{
  \item{\code{bogus}}{a character vector}
  \item{\code{Timestamp}}{Timestamp indicates when the data were entered into an electronic record,
not anything about the bird being described}
  \item{\code{Year}}{a character vector}
  \item{\code{Day}}{a character vector}
  \item{\code{Month}}{a character vector}
 \item{\code{CaptureTime}}{a character vector}
 \item{\code{SpeciesName}}{a character vector}
 \item{\code{Sex}}{a character vector}
 \item{\code{Age}}{a character vector}
 \item{\code{BandNumber}}{a character vector}
 \item{\code{TrapID}}{a character vector}
 \item{\code{Weather}}{a character vector}
 \item{\code{BandingReport}}{a character vector}
 \item{\code{RecaptureYN}}{a character vector}
 \item{\code{RecaptureMonth}}{a character vector}
 \item{\code{RecaptureDay}}{a character vector}
 \item{\code{Condition}}{a character vector}
 \item{\code{Release}}{a character vector}
 \item{\code{Comments}}{a character vector}
 \item{\code{DataEntryPerson}}{a character vector}
 \item{\code{Weight}}{a character vector}
 \item{\code{WingChord}}{a character vector}
 \item{\code{Temperature}}{a character vector}
 \item{\code{RecaptureOriginal}}{a character vector}
 \item{\code{RecapturePrevious}}{a character vector}
 \item{\code{TailLength}}{a character vector}
 }

Timestamp indicates when the data were entered into an electronic record,
not anything about the bird being described.}
\source{
Jerald Dosch, Dept. of Biology, Macalester College: the manager of
the Study Area.
}
\usage{
OrdwayBirds
}
\description{
The historical record of birds captured and released at the Katharine Ordway
Natural History Study Area, a 278-acre preserve in Inver Grove Heights,
Minnesota, owned and managed by Macalester College.
}
\details{
There are many extraneous levels of variables such as species.  Part of the
purpose of this data set is to teach about data cleaning.
}
\examples{

data(OrdwayBirds)

}
\keyword{datasets}
