% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAT_2010}
\alias{SAT_2010}
\title{State SAT scores from 2010}
\format{
A data.frame with 50 rows and 9 variables.
\describe{
\item{state}{a factor with levels for each state}
\item{expenditure}{average expenditure per student (in each state)}
\item{pupil_teacher_ratio}{pupil to teacher ratio in that state}
\item{salary}{teacher salary (in 2010 US $)}
\item{read}{state average Reading SAT score}
\item{math}{state average Math SAT score}
\item{write}{state average Writing SAT score}
\item{total}{state average Total SAT score}
\item{sat_pct}{percent of students taking SAT in that state}
}
}
\usage{
SAT_2010
}
\description{
SAT results by state for 2010
}
\details{
See also the earlier \link[mosaicData:SAT]{mosaicData::SAT} dataset.
}
\seealso{
\link[mosaicData:SAT]{mosaicData::SAT}
}
\keyword{datasets}
