% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAT_2010}
\alias{SAT_2010}
\title{State SAT scores form 2010}
\format{A data.frame with 50 rows and 9 variables.}
\usage{
SAT_2010
}
\description{
SAT results by state for 2010
}
\keyword{datasets}

