% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scidb.R
\name{src_scidb}
\alias{dbConnect_scidb}
\alias{src_scidb}
\title{src_scidb}
\usage{
src_scidb(dbname, ...)

dbConnect_scidb(dbname, ...)
}
\arguments{
\item{dbname}{the name of the database to which you want to connect}

\item{...}{arguments passed to \code{\link[dplyr]{src_mysql}} or \code{\link[DBI]{dbConnect}}}
}
\value{
For \code{\link{src_mysql}}, a \code{\link[dplyr]{src_mysql}} object

For \code{\link{dbConnect_scidb}}, a \code{\link[DBI]{DBIConnection-class}} object
}
\description{
Connect to the scidb server at Smith

dbConnect_scidb
}
\details{
this is a public, read-only account. Any abuse will be considered a
hostile act.
}
\examples{
dbAir <- src_scidb("airlines")
dbAir
dbAir <- dbConnect_scidb("airlines")
dbAir
}
\seealso{
\code{\link[dplyr]{src_mysql}}

\code{\link[DBI]{DBIConnection-class}}
}

