% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Votes}
\alias{Votes}
\title{Votes from Scottish Parliament}
\format{An object of class \code{data.frame} with 774 rows and 135 columns.}
\usage{
Votes
}
\description{
Votes from Scottish Parliament
}
\keyword{datasets}

