% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldCities}
\alias{WorldCities}
\title{Cities and their populations}
\format{
A data frame with 23,018 observations on the following 10 variables.
\describe{ 
\item{code}{The ISO (?) city code}
\item{name}{Name of the city} 
\item{latitude}{location in degrees} 
\item{longitude}{location in degrees}
\item{country}{Two letter country code}
\item{countryRegion}{A numerical region}
\item{population}{Population} 
\item{regionCode}{ISO (?) Code} 
\item{region}{Name of the region} 
\item{date}{Date estimate made} }
}
\source{
In Draft: Somewhere on the Internet.  We need a proper source.
}
\usage{
WorldCities
}
\description{
A list of cities
}
\examples{

data(WorldCities)

}
\references{
In Draft: We need a proper reference
}
\keyword{datasets}
